requirements.txt
astroid==2.4.2
autopep8==1.5.3
backcall==0.2.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cycler==0.10.0
decorator==4.4.2
Flask==1.1.2
Flask-HTTPAuth==4.1.0
idna==2.10
imageio==2.9.0
imutils==0.5.3
ipykernel==5.3.4
ipython==7.17.0
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
jupyter-client==6.1.6
jupyter-core==4.6.3
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.2.2
mccabe==0.6.1
networkx==2.4
numpy==1.19.0
opencv-contrib-python==3.4.2.16
opencv-python==3.4.2.16
parso==0.7.1
pickleshare==0.7.5
Pillow==7.2.0
prompt-toolkit==3.0.5
pycodestyle==2.6.0
Pygments==2.6.1
pylint==2.5.3
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
pywin32==228
pyzmq==19.0.2
requests==2.24.0
scikit-image==0.17.2
scipy==1.5.1
six==1.15.0
tifffile==2020.7.4
toml==0.10.1
tornado==6.0.4
traitlets==4.3.3
typed-ast==1.4.1
urllib3==1.25.10
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
